Unit tests for routes and register
