--- test_basic.py
@@ -1,28 +1,0 @@
-import os
-import unittest
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_bcrypt import Bcrypt
-from flask_login import LoginManager
- 
-from PotHoleReporter import application as app, db, loginManager
- 
-class BasicTests(unittest.TestCase):
-    def setUp(self):
-        app.config['TESTING'] = True
-        app.config['WTF_CSRF_ENABLED'] = False
-        app.config['DEBUG'] = False
-        self.app = app.test_client()
- 
-    def tearDown(self):
-        pass
-
-    def test_main_page(self):
-        response = self.app.get('/', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        print("TEST")
- 
- 
-if __name__ == "__main__":
-    unittest.main()
-    #BasicTests()
